+++ CHANGELOG.rst
@@ -1,6 +1,7 @@
 **0.1.2 - TBD**
 
  - Add mypy type checking
+ - Add unit tests for FuzzyChecker
 
 **0.1.1 - 10/14/24**